Use specific classes in new BiMap and move it to driver package.
